@@ -9,10 +9,11 @@
     <version>1.0-SNAPSHOT</version>
 
     <properties>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
-        <springboot.version>2.6.6</springboot.version>
-        <satoken.version>1.33.0</satoken.version>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
+        <java.version>17</java.version>
+        <springboot.version>3.0.2</springboot.version>
+        <satoken.version>1.34.0</satoken.version>
         <common-pool.version>2.11.1</common-pool.version>
         <lombok.version>1.18.26</lombok.version>
         <commons.lang.version>3.12.0</commons.lang.version>
@@ -51,7 +52,7 @@
         <!-- auth server -->
         <dependency>
             <groupId>cn.dev33</groupId>
-            <artifactId>sa-token-spring-boot-starter</artifactId>
+            <artifactId>sa-token-spring-boot3-starter</artifactId>
             <version>${satoken.version}</version>
         </dependency>
 
