@@ -2,7 +2,7 @@ package com.zjuici.authserver.dao.po;
 
 import lombok.Data;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.util.Date;
 
 /**
