@@ -1,8 +1,8 @@
 package com.zjuici.authserver.dao.po;
 
+import jakarta.persistence.*;
 import lombok.Data;
 
-import javax.persistence.*;
 import java.io.Serializable;
 import java.util.Date;
 
