@@ -3,10 +3,10 @@ package com.zjuici.authserver.service.impl;
 import com.zjuici.authserver.dao.po.UpmResource;
 import com.zjuici.authserver.dao.repository.UpmResourceRepository;
 import com.zjuici.authserver.service.UpmResourceService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Service;
 
-import javax.annotation.Resource;
 
 /**
  * @ClassName UpmResourceServiceImpl
@@ -22,7 +22,7 @@ public class UpmResourceServiceImpl implements UpmResourceService {
     /**
      * upmResourceRepository
      */
-    @Resource
+    @Autowired
     private UpmResourceRepository upmResourceRepository;
 
     @Override
